# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "get": {
        "operationId": "ProductsController_findAll_v1",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "section",
            "required": false,
            "in": "query",
            "description": "Section filter (best_deals, recommended, top_sellers, new_arrivals)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Category slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subCategory",
            "required": false,
            "in": "query",
            "description": "Sub-category slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "brand",
            "required": false,
            "in": "query",
            "description": "Single brand name (one brand filter at a time)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "Minimum price",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Maximum price",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "inStock",
            "required": false,
            "in": "query",
            "description": "In stock only",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "ratingGte",
            "required": false,
            "in": "query",
            "description": "Minimum rating",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "relevance",
              "type": "string",
              "enum": [
                "relevance",
                "price_asc",
                "price_desc",
                "newest",
                "popular"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "items": [
                      {
                        "id": "6710def123abc456789012ab",
                        "slug": "hisense-43-inch-smart-tv",
                        "name": "Hisense 43\" Smart TV",
                        "brand": "Hisense",
                        "thumbnailUrl": "https://res.cloudinary.com/.../hisense-43.jpg",
                        "price": {
                          "amount": 220000,
                          "currency": "NGN",
                          "formatted": "₦220,000"
                        },
                        "originalPrice": {
                          "amount": 260000,
                          "currency": "NGN",
                          "formatted": "₦260,000"
                        },
                        "savings": {
                          "amount": 40000,
                          "currency": "NGN",
                          "formatted": "₦40,000"
                        },
                        "discountPercent": 15,
                        "rating": 4.6,
                        "reviewCount": 184,
                        "inventoryStatus": "in_stock",
                        "badge": "Best Deal",
                        "seller": {
                          "id": "6710abc...",
                          "name": "AlphaTech Store",
                          "isVerified": true,
                          "location": "Lagos",
                          "deliveryTimeRange": "1-3 days"
                        }
                      }
                    ],
                    "pagination": {
                      "page": 1,
                      "limit": 10,
                      "totalItems": 142,
                      "totalPages": 15,
                      "hasNextPage": true,
                      "hasPrevPage": false
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "List products with search, filter, and pagination",
        "tags": [
          "Catalogue"
        ]
      },
      "post": {
        "operationId": "ProductsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a product (Seller or Admin)",
        "tags": [
          "Catalogue"
        ]
      }
    },
    "/products/{idOrSlug}": {
      "get": {
        "operationId": "ProductsController_findOne_v1",
        "parameters": [
          {
            "name": "idOrSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the full product detail including variant axes, gallery, breadcrumbs, specs, tabs and seller profile."
          }
        },
        "summary": "Get product detail by Mongo id or slug",
        "tags": [
          "Catalogue"
        ]
      }
    },
    "/products/{productId}/related": {
      "get": {
        "operationId": "ProductsController_findRelated_v1",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get related products",
        "tags": [
          "Catalogue"
        ]
      }
    },
    "/products/{productId}/variants": {
      "get": {
        "operationId": "ProductsController_findVariants_v1",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get product variants and axes",
        "tags": [
          "Catalogue"
        ]
      }
    },
    "/products/{productId}/availability": {
      "get": {
        "operationId": "ProductsController_findAvailability_v1",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get product/variant stock availability",
        "tags": [
          "Catalogue"
        ]
      }
    },
    "/products/{productId}/policies": {
      "get": {
        "operationId": "ProductsController_findPolicies_v1",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get product return and warranty policies",
        "tags": [
          "Catalogue"
        ]
      }
    },
    "/products/{id}": {
      "patch": {
        "operationId": "ProductsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a product (owner Seller or Admin)",
        "tags": [
          "Catalogue"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a product (owner Seller or Admin)",
        "tags": [
          "Catalogue"
        ]
      }
    },
...
    "/policies/products/{productId}": {
      "get": {
        "operationId": "PoliciesController_getProductPolicy_v1",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get product return and warranty policies",
        "tags": [
          "Catalogue"
        ]
      }
    },
...
    "/sellers/me/products/{id}": {
      "patch": {
        "operationId": "SellersController_updateMyProduct_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update one of your own product listings",
        "tags": [
          "Sellers"
        ]
      },
      "delete": {
        "operationId": "SellersController_removeMyProduct_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Withdraw one of your own product listings",
        "tags": [
          "Sellers"
        ]
      }
    },
...
    "/products/{productId}/reviews": {
      "get": {
        "operationId": "ReviewsController_findByProduct_v1",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get product reviews with summary",
        "tags": [
          "Catalogue"
        ]
      },
      "post": {
        "operationId": "ReviewsController_createReview_v1",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Write a review for a product (one per user per product)",
        "tags": [
          "Catalogue"
        ]
      }
    },
    "/products/{productId}/reviews/{reviewId}/helpful": {
      "post": {
        "operationId": "ReviewsController_markHelpful_v1",
        "parameters": [
          {
            "name": "reviewId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mark a review as helpful (no auth required)",
        "tags": [
          "Catalogue"
        ]
      }
    },
...
    "/admin/products/{id}": {
      "patch": {
        "operationId": "AdminProductsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update any product",
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "AdminProductsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Force delete any product",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cadna Mart API",
    "description": "Backend API for the Cadna Mart marketplace.\n\n### Response envelope\nEvery successful response is wrapped by the global response interceptor:\n```json\n{\n  \"success\": true,\n  \"statusCode\": 200,\n  \"data\": { ... },\n  \"message\": \"...\",\n  \"meta\": { \"correlationId\": \"...\", \"timestamp\": \"...\" }\n}\n```\nWhen documenting individual endpoints below, the schema describes the `data` payload only.\n\n### Authentication\n- **Bearer JWT** (`Authorization: Bearer <accessToken>`) for authenticated user routes.\n- **Guest token** (`x-guest-token: <token>`) for guest-cart endpoints. Returned once from `POST /cart`.\n\n### Error shape\nErrors flow through the global `AllExceptionsFilter`:\n```json\n{\n  \"success\": false,\n  \"statusCode\": 400,\n  \"message\": \"Validation failed\",\n  \"errorCode\": \"BAD_REQUEST\",\n  \"path\": \"/api/v1/...\",\n  \"correlationId\": \"...\"\n}\n```",
    "version": "1.0",
    "contact": {
      "name": "Cadna Mart Engineering",
      "url": "",
      "email": "cadnamart@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "https://cadna-mart-be-nsz2.onrender.com/api/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token returned from /auth/login or /auth/otp/verify"
      },
      "guestToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-guest-token",
        "description": "Guest cart access token (returned once from POST /cart for guests)"
      }
    },
    "schemas": {
...
      "GalleryImageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "alt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "url"
        ]
      },
      "VariantOptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "swatchHex": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "label"
        ]
      },
      "VariantAxisDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantOptionDto"
            }
          }
        },
        "required": [
          "name",
          "displayName",
          "options"
        ]
      },
      "ProductVariantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "attributes": {
            "type": "object"
          },
          "priceAmount": {
            "type": "number"
          },
          "stockQty": {
            "type": "number"
          },
          "isInStock": {
            "type": "boolean"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "sku",
          "attributes",
          "priceAmount"
        ]
      },
      "ProductTabDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "contentHtml": {
            "type": "string"
          }
        },
        "required": [
          "label",
          "contentHtml"
        ]
      },
      "SpecificationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "value"
        ]
      },
      "BreadcrumbDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "url": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "label"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "sellerId": {
            "type": "string",
            "description": "Seller ID this product belongs to"
          },
          "categoryId": {
            "type": "string"
          },
          "subCategoryId": {
            "type": "string"
          },
          "priceAmount": {
            "type": "number",
            "description": "Selling price in NGN (kobo or naira depending on convention)"
          },
          "originalPriceAmount": {
            "type": "number",
            "description": "Original/compare-at price (if on sale)"
          },
          "discountPercent": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "gallery": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GalleryImageDto"
            }
          },
          "variantAxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantAxisDto"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantDto"
            }
          },
          "defaultVariantId": {
            "type": "string"
          },
          "descriptionHtml": {
            "type": "string"
          },
          "tabs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductTabDto"
            }
          },
          "specifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpecificationDto"
            }
          },
          "breadcrumbs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BreadcrumbDto"
            }
          },
          "badge": {
            "type": "string"
          },
          "sections": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean"
          },
          "inventoryStatus": {
            "type": "string"
          },
          "returnPolicy": {
            "type": "object",
            "description": "Return policy for this product"
          }
        },
        "required": [
          "name",
          "sellerId",
          "priceAmount"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "sellerId": {
            "type": "string",
            "description": "Seller ID this product belongs to"
          },
          "categoryId": {
            "type": "string"
          },
          "subCategoryId": {
            "type": "string"
          },
          "priceAmount": {
            "type": "number",
            "description": "Selling price in NGN (kobo or naira depending on convention)"
          },
          "originalPriceAmount": {
            "type": "number",
            "description": "Original/compare-at price (if on sale)"
          },
          "discountPercent": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "gallery": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GalleryImageDto"
            }
          },
          "variantAxes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantAxisDto"
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantDto"
            }
          },
          "defaultVariantId": {
            "type": "string"
          },
          "descriptionHtml": {
            "type": "string"
          },
          "tabs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductTabDto"
            }
          },
          "specifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpecificationDto"
            }
          },
          "breadcrumbs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BreadcrumbDto"
            }
          },
          "badge": {
            "type": "string"
          },
          "sections": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean"
          },
          "inventoryStatus": {
            "type": "string"
          },
          "returnPolicy": {
            "type": "object",
            "description": "Return policy for this product"
          }
        }
      },
...
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5
          },
          "title": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "rating"
        ]
      },
...
    }
  }
}